requirements/prod.txt
-r common.txt

django-pylibmc==0.6.1
pylibmc==1.5.0
raven==5.9.2